requirements.txt
streamlit
gspread
oauth2client
pandas
google-genai